課題４修正4

+++ exercise4/practice4.c
@@ -7,8 +7,8 @@ int main(void)
 
     p = &i;
 
-    printf("i: アドレス=%#llx 値=%d\n", (unsigned long)&i, i);
-    printf("p: アドレス=%#llx 値=%#llx 指す先の値=%d\n", (unsigned long)&p, (unsigned long)p, *p);
+    printf("i: アドレス=%#llx 値=%d\n", &i, i);
+    printf("p: アドレス=%#llx 値=%#llx 指す先の値=%d\n", &p, p, *p);
 
     return 0;
 }
